# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Skolara API",
    "description": "Backend for searching colleges, managing applications, and processing payments.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8000/api" },
    { "url": "https://devapi.skolara.in/api" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/user/profile": {
      "get": {
        "tags": ["User"],
        "summary": "Get User Profile",
        "operationId": "getUserProfile",
        "responses": {
          "200": {
            "description": "Profile fetched successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetUserResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "403": { "$ref": "#/components/responses/JWTForbidden" }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "Update User Profile",
        "operationId": "updateUserProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              },
              "examples": {
                "valid": {
                  "value": {
                    "first_name": "User",
                    "last_name": "Skolara",
                    "email": "user@example.com",
                    "mobile": "9876543210",
                    "gender": "Male",
                    "dob": "2001-12-31"
                  }
                },
                "nullables": {
                  "value": {
                    "first_name": "User",
                    "last_name": "Skolara",
                    "email": "user@example.com",
                    "mobile": "9876543210",
                    "gender": null,
                    "dob": null
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserProfileResponse"
                },
                "examples": {
                  "ok": {
                    "value": { "message": "Profile updated successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or email/mobile conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "conflict": {
                    "value": { "error": "Email or Mobile already exists" }
                  },
                  "invalid": { "value": { "error": "INVALID_REQUEST" } }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "403": { "$ref": "#/components/responses/JWTForbidden" },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "notFound": { "value": { "error": "User not found" } }
                }
              }
            }
          }
        }
      }
    },
...
    "/user/notifications/all": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List notifications for the authenticated user",
        "operationId": "getNotifications",
        "responses": {
          "200": {
            "description": "Notifications (newest first)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "notFound": { "value": { "error": "User not found" } }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/user/notifications/add": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Create a notification for the authenticated user",
        "operationId": "addNotification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NotificationPayload" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                },
                "examples": {
                  "ok": {
                    "value": { "message": "Notification added successfully" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "notFound": { "value": { "error": "User not found" } }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/user/notifications/mark-read/{id}": {
      "put": {
        "tags": ["Notifications"],
        "summary": "Mark a notification as read",
        "operationId": "markNotificationAsRead",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                },
                "examples": {
                  "ok": {
                    "value": { "message": "Notification marked as read" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "Notification not found or not owned by user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "nf": { "value": { "error": "Notification not found" } }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/user/support": {
      "put": {
        "tags": ["Support & FAQs"],
        "summary": "Record a support request for a complaint",
        "operationId": "recordSupport",
        "parameters": [
          {
            "in": "query",
            "name": "complaint_id",
            "required": true,
            "schema": { "type": "number" },
            "description": "Complaint ID to associate"
          }
        ],
        "responses": {
          "201": {
            "description": "Support recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                },
                "examples": {
                  "ok": {
                    "value": {
                      "message": "Support request recorded successfully"
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "User or Complaint not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "user": { "value": { "error": "User not found" } },
                  "complaint": { "value": { "error": "Complaint not found" } }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Standard JWT bearer token (used on protected endpoints)."
      }
    },
    "responses": {
      "JWTUnauthorized": {
        "description": "Missing, malformed, expired, or invalid JWT",
        "headers": {
          "WWW-Authenticate": {
            "schema": { "type": "string" },
            "description": "Auth challenge header (e.g., Bearer realm and error details)"
          }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/JWTErrorResponse" },
            "examples": {
              "missingHeader": {
                "value": {
                  "error": "UNAUTHORIZED",
                  "message": "Missing or malformed Authorization header",
                  "statusCode": 401
                }
              },
              "invalid": {
                "value": {
                  "error": "INVALID_TOKEN",
                  "message": "JWT signature verification failed",
                  "statusCode": 401
                }
              },
              "expired": {
                "value": {
                  "error": "TOKEN_EXPIRED",
                  "message": "JWT access token has expired",
                  "statusCode": 401
                }
              }
            }
          }
        }
      },
      "JWTForbidden": {
        "description": "Authenticated but not allowed to access this resource",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/JWTErrorResponse" },
            "examples": {
              "forbidden": {
                "value": {
                  "error": "FORBIDDEN",
                  "message": "Insufficient permissions",
                  "statusCode": 403
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "mobile": { "type": "string", "pattern": "^[0-9]{10}$" },
          "email": { "type": "string", "format": "email", "nullable": true },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "gender": {
            "type": "string",
            "enum": ["Male", "Female", "Other"],
            "nullable": true
          },
          "dob": { "type": "string", "format": "date", "nullable": true },
          "prefered_theme": { "type": "string", "enum": ["Light", "Dark"] },
          "allow_notifications": { "type": "boolean" }
        },
        "required": [
          "user_id",
          "mobile",
          "email",
          "first_name",
          "last_name",
          "prefered_theme",
          "allow_notifications"
        ]
      },
...
      "JWTErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "UNAUTHORIZED",
              "INVALID_TOKEN",
              "TOKEN_EXPIRED",
              "FORBIDDEN"
            ],
            "description": "JWT validation error type"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error details"
          },
          "statusCode": {
            "type": "integer",
            "example": 401,
            "description": "HTTP status code"
          }
        },
        "required": ["error"]
      },
      "UpdateUserProfileRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["first_name", "last_name", "email", "mobile"],
        "properties": {
          "first_name": { "type": "string", "minLength": 1 },
          "last_name": { "type": "string", "minLength": 1 },
          "email": { "type": "string", "format": "email" },
          "mobile": { "type": "string", "pattern": "^[0-9]{10}$" },
          "gender": {
            "type": "string",
            "enum": ["Male", "Female", "Others"],
            "nullable": true,
            "default": null
          },
          "dob": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "default": null,
            "description": "YYYY-MM-DD"
          }
        }
      },
      "UpdateUserProfileResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Profile updated successfully"
          }
        },
        "required": ["message"]
      },
...
      "NotificationPayload": {
        "type": "object",
        "required": ["title", "subtitle", "notificationId"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "subtitle": { "type": "string", "minLength": 1, "maxLength": 255 },
          "notificationId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "imageUrl": {
            "type": ["string", "null"],
            "maxLength": 1000,
            "nullable": true
          },
          "navigationUrl": {
            "type": ["string", "null"],
            "maxLength": 1000,
            "nullable": true
          },
          "priority": {
            "type": "string",
            "enum": ["High", "Medium", "Low"],
            "default": "Medium",
            "nullable": true
          }
        }
      },
      "NotificationItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "subtitle": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "is_read": { "type": "boolean" },
          "imageUrl": {
            "type": ["string", "null"],
            "format": "uri",
            "nullable": true
          },
          "navigationUrl": {
            "type": ["string", "null"],
            "format": "uri",
            "nullable": true
          },
          "priority": {
            "type": ["string", "null"],
            "enum": ["High", "Medium", "Low"],
            "nullable": true
          },
          "notificationId": { "type": "string" }
        },
        "required": [
          "id",
          "title",
          "subtitle",
          "created_at",
          "is_read",
          "notificationId"
        ]
      },
      "NotificationListResponse": {
        "type": "array",
        "items": { "$ref": "#/components/schemas/NotificationItem" }
      },
      "GenericMessageResponse": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"]
      },
...
    }
  }
}